requirements.txt
langchain==0.0.150
InstructorEmbedding==1.0.1
faiss-cpu>=1.7.4,<2.0
openai==0.28.1
streamlit==1.40.2
PyPDF2==3.0.1
python-dotenv==1.0.0
tiktoken==0.8.0
transformers==4.28.1
sentence-transformers==2.2.2
huggingface_hub==0.11.1  # Ensure compatibility with InstructorEmbedding
torch==1.13.1  # Add PyTorch for InstructorEmbedding
pymupdf>=1.20.2,<2.0
